Refactor: Blocks with age in BlockState use BlockAgeTrait
