+++ src/block/CaveVinesHeadWithBerries.php
@@ -25,8 +25,8 @@ class CaveVinesHeadWithBerries extends BaseCaveVines{
 	}
 
 	public function getStateDeserialize() : ?Closure{
-		return static fn(BlockStateReader $in) : CaveVinesHeadWithBerries => ExtraVanillaBlocks::CAVE_VINES_HEAD_WITH_BERRIES()
-			->setAge($in->readBoundedInt(BlockStateNames::GROWING_PLANT_AGE, 0, self::MAX_AGE));
+		return fn(BlockStateReader $in) : CaveVinesHeadWithBerries => ExtraVanillaBlocks::CAVE_VINES_HEAD_WITH_BERRIES()
+			->setAge($in->readBoundedInt(BlockStateNames::GROWING_PLANT_AGE, 0, $this->getMaxAge()));
 	}
 
 	public function onInteract(Item $item, int $face, Vector3 $clickVector, ?Player $player = null, array &$returnedItems = []) : bool{
