+++ src/block/Kelp.php
@@ -15,16 +15,16 @@ use pocketmine\data\runtime\RuntimeDataReader;
 use pocketmine\data\runtime\RuntimeDataWriter;
 use pocketmine\item\Item;
 use pocketmine\math\Facing;
+use skh6075\pmexpansion\block\utils\BlockAgeTrait;
 use skh6075\pmexpansion\block\utils\BlockTypeIdTrait;
 use skh6075\pmexpansion\block\utils\IBlockState;
 use skh6075\pmexpansion\item\ExtraVanillaItems;
 
 class Kelp extends Transparent implements IBlockState{
 	use BlockTypeIdTrait;
+	use BlockAgeTrait;
 
-	protected const MAX_AGE = 25;
-
-	protected int $age = 0;
+	public function getMaxAge() : int{ return 25; }
 
 	public function getStateSerialize() : ?Closure{
 		return static fn(Kelp $block) : BlockStateWriter => BlockStateWriter::create(BlockTypeNames::KELP)
@@ -32,24 +32,14 @@ class Kelp extends Transparent implements IBlockState{
 	}
 
 	public function getStateDeserialize() : ?Closure{
-		return static fn(BlockStateReader $in) : Kelp => ExtraVanillaBlocks::KELP()
-			->setAge($in->readBoundedInt(BlockStateNames::KELP_AGE, 0, self::MAX_AGE));
+		return fn(BlockStateReader $in) : Kelp => ExtraVanillaBlocks::KELP()
+			->setAge($in->readBoundedInt(BlockStateNames::KELP_AGE, 0, $this->getMaxAge()));
 	}
 
 	public function getRequiredStateDataBits() : int{ return 5; }
 
 	protected function describeState(RuntimeDataWriter|RuntimeDataReader $w) : void{
-		$w->boundedInt(5, 0, self::MAX_AGE, $this->age);
-	}
-
-	public function getAge(): int{ return $this->age; }
-
-	public function setAge(int $age): self{
-		if($age < 0 || $age > self::MAX_AGE){
-			throw new \InvalidArgumentException("Age must be in range 0 ... " . self::MAX_AGE);
-		}
-		$this->age = $age;
-		return $this;
+		$w->boundedInt(5, 0, $this->getMaxAge(), $this->age);
 	}
 
 	protected function recalculateCollisionBoxes() : array{ return []; }
